InfoMagnus-master/dsdp/requirements.txt
amqp==2.5.1
billiard==3.6.1.0
boto3==1.9.236
botocore==1.12.236
certifi==2019.9.11
chardet==3.0.4
Click==7.0
configobj==5.0.6
dnslib==0.9.10
dnspython==1.16.0
docopt==0.6.2
docutils==0.15.2
Flask==1.1.1
Flask-Classful==0.14.2
Flask-Middleware==1.0
idna==2.8
importlib-metadata==0.23
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.9.4
kombu==4.6.3
MarkupSafe==1.1.1
more-itertools==7.2.0
pandas==0.25.2
psycopg2-binary==2.8.3
pyaes==1.6.0
python-dateutil==2.8.0
pytz==2019.2
requests==2.22.0
s3transfer==0.2.1
six==1.12.0
subprocess32==3.5.4
tzlocal==2.0.0
urllib3==1.25.6
vine==1.3.0
Werkzeug==0.16.0
zipp==0.6.0
uwsgi==2.0.17
celery==4.3.0
redis==3.3.11
flower==0.9.3
eventlet==0.25.1
tensorflow==1.13.1
paramiko==2.6.0
SQLAlchemy==1.3.11
marshmallow==3.0.3